+++ human-verification/data/src/main/kotlin/me/proton/core/humanverification/data/db/HumanVerificationConverters.kt
@@ -31,4 +31,13 @@ class HumanVerificationConverters {
     fun fromStringToHumanVerificationState(value: String?): HumanVerificationState? = value?.let {
         HumanVerificationState.valueOf(value)
     }
+
+    @TypeConverter
+    fun fromClientIdTypeToString(value: ClientIdType?): String? = value?.value?.lowercase()
+
+    @TypeConverter
+    fun fromStringToClientIdType(value: String?): ClientIdType? = value?.let {
+        ClientIdType.map[value.lowercase()]
+    }
+
 }